Add countdown for task to update top artists on auth create
There is a race condition due to the usage of a transaction block
during the creation of a SpotifyUserAuth record. This caused the
task to be fired from the signal before the instance was saved to
the database, which caused the task to fail in the lookup of the
auth record. Adding a countdown before calling the task should
ensure that the task gets called after the record is saved and
inserted into the table.