Fix regex for building first and last mood playlist page links
Originally the regex was missing the 0 character, which caused the
links to erroneously include the 0 char in the query string. This
updates the regex to also strip the 0 char from the params to
ensure that the anchor links are built properly.

Also add more tests around the params for the anchor links.