Redirect password change success requests to profile page
Change wording of message for password change

Now that we redirect authenticated users from the login page,
password change successes are redirected to the default login
redirect on requests to login page. Update the redirect URL
for password change success to where users would be
redirected if they were authenticated and update the wording
of the message.