fix: properly randomize browse emotion if not set
The original implementation did not randomize the emotion if the
initial value was not passed. This resulted in Melancholy being
displayed every time, which is not ideal as we want the emotion to
be random to encourage users to browse across all the provided
emotions in our system.

This adds a check to see if we have an initial emotion set, and if
we do not then pick a random one from the available choices.