feat: hide Spotify play button border
It looks like Spotify updated the styling of their play buttons
to be more rounded, which results in a weird empty border for our
iframes when we load a play button. This change hides the border
for these iframes so there isn't a weird box looking outline around
the otherwise rounded play buttons.